TPAStar initial
TPAStar is working but the surface path projection will break when
consecutive corners need to be rounded because it needs to search
all tris connected to the corner point.

The issue with this is that to get TPAStar to work we deleted the
tris that are creating the obstacle that makes the corner, so we
can't walk around a point using halfedges as some of those
tris/edges no longer exist.

To fix this we'll need to get rid of the deep-clean of impassable
tris and instead allow TPAStar to store a set of obstacle tris