TPAStar Surface fixes and optimal path fix
The surface projection was getting stuck when navigating around
obstacles with multiple points, since it would move to an exact
point on the height mesh and then not know which edge to cross
to the next triangle.

The fix for this is to rotate around the point testing each
triangle to see if it can handle the next path segment, but
this meant removing the deep tri removal since in order
to find triangles connected to a point within delaunator it
needs to be able to wind around half edges, so there can't be
gaps in the triangulation.

To maintain obstacles without deleting their triangles a new struct
storing flags for each tri has been added which can optionally be
provided to a triangulation, and TPA will respect this whenever
it fetches the neighbours of a triangle.

In addition, when adding a path to the Open Set, TPAStar was adding
paths that should have been at the end to the beginning, resulting
in some particularly scenic routes.